+++ nanoc/test/data_sources/test_filesystem.rb
@@ -145,10 +145,10 @@ class Nanoc::DataSources::FilesystemTest < Nanoc::TestCase
 
     # Get input and expected output
     expected = {
-      '/foo' => Nanoc::Identifier.new('/foo', type: :full),
-      '/foo.html' => Nanoc::Identifier.new('/foo.html',       type: :full),
-      '/foo/index.html' => Nanoc::Identifier.new('/foo/index.html', type: :full),
-      '/foo.html.erb' => Nanoc::Identifier.new('/foo.html.erb', type: :full),
+      '/foo' => Nanoc::Core::Identifier.new('/foo', type: :full),
+      '/foo.html' => Nanoc::Core::Identifier.new('/foo.html',       type: :full),
+      '/foo/index.html' => Nanoc::Core::Identifier.new('/foo/index.html', type: :full),
+      '/foo.html.erb' => Nanoc::Core::Identifier.new('/foo.html.erb', type: :full),
     }
 
     # Check